split the GC integration into diagnostics_gc.rs — diagnostics.rs must stay under the 2000-line gate
